requirement.txt
numpy==1.18.5
pandas==1.1.4
tensorflow-gpu==2.3.1
matplotlib==3.3.2
scikit-learn==0.23.2